+++ src/db.rs
@@ -1,1 +1,2 @@
 pub mod connect;
+pub mod email;
